\name{E_aguiar}
\alias{aguiar}
\alias{MTM}
\alias{Ktlim}
\alias{Ktm}
\docType{data}
\title{Markov Transition Matrices for the Aguiar etal. procedure}
\description{Markov Transition Matrices and auxiliary data for generating
  sequences of daily radiation values.}

\usage{data(MTM)}

\format{
  \code{MTM} is a \code{data.frame} with the collection of Markov
  Transition Matrices defined in the paper "Simple procedure for
  generating sequences of daily radiation values using a library of
  Markov transition matrices", Aguiar et al., Solar Energy,
  1998. \code{Ktlim} (matrix) and \code{Ktm} (vector) are auxiliary data
  to choose the correspondent matrix of the collection.
}

\keyword{datasets}
