\name{solaR-defunct}
\encoding{UTF-8}
\alias{solaR-defunct}
\alias{readSIAR}

\title{Defunct functions in package \sQuote{solaR}}

\description{
  These functions are defunct and no longer available.
}

\details{
\code{readSIAR}: The SIAR webpage cannot be accessed with a direct
      URL but using javascript code. Therefore, the function
      \code{readSIAR} no longer works. This help page is still here as a
      reference. The SIAR webpage is now \url{http://eportal.magrama.gob.es/websiar}.
}
