% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_irradiance.R
\name{get_irradiance}
\alias{get_irradiance}
\title{Get the irradiance for a given location and time}
\usage{
get_irradiance(
  year = "2023",
  month = "1",
  day = "1",
  lat = "-34.2833",
  lon = "150.95",
  min_wavelength = "300",
  max_wavelength = "560",
  wavelength_interval = "10"
)
}
\arguments{
\item{year}{The year for which the irradiance is requested.}

\item{month}{The month for which the irradiance is requested.}

\item{day}{The day for which the irradiance is requested.}

\item{lat}{The latitude for which the irradiance is requested.}

\item{lon}{The longitude for which the irradiance is requested.}

\item{min_wavelength}{The minimum wavelength for which the irradiance is requested.}

\item{max_wavelength}{The maximum wavelength for which the irradiance is requested.}

\item{wavelength_interval}{The wavelength interval for which the irradiance is requested.}
}
\value{
A tibble with the wavelength, category, and value of the irradiance.
}
\description{
Get the irradiance for a given location and time
}
\examples{
get_irradiance()
}
