\name{D_getData-methods}
\docType{methods}
\alias{getData}
\alias{getData-methods}
\alias{getData,Meteo-method}
\title{Methods for function getData}
\description{
  Meteorological source data of a \code{Meteo} (or extended) object.
}
\section{Methods}{
  \describe{

    \item{\code{signature(object = "Meteo")}}{returns the meteorological
      source data of the slot \code{data} of the object.
    }
}}
\keyword{methods}
\author{Oscar Perpiñán Lamigueiro}
