\name{D_getG0-methods}
\docType{methods}
\alias{getG0}
\alias{getG0-methods}
\alias{getG0,Meteo-method}
\title{Methods for function getG0}
\description{
  Global irradiation source data of a \code{Meteo} (or extended) object.
}
\section{Methods}{
  \describe{

    \item{\code{signature(object = "Meteo")}}{returns the global irradiation
      values stored in a \code{Meteo} object.
    }
}}
\keyword{methods}
\author{Oscar Perpiñán Lamigueiro}