\name{D_indexRep-methods}
\docType{methods}
\alias{indexRep-methods}
\alias{indexRep,Sol-method}
\title{Methods for Function indexRep}
\description{Daily time index of \code{solaR} object.}
\section{Methods}{
  \describe{
    \item{\code{signature(object = "Sol")}}{returns the daily index of
      the \code{solD} slot but repeated to match the length of the index of the
      \code{solI} slot.
    }
}}
\keyword{methods}
\author{Oscar Perpiñán Lamigueiro}
