\name{D_plot-methods}
\docType{methods}
\alias{plot-methods}
\alias{plot,Shade-method}
\title{Methods for Function plot in Package `solaR'}
\description{
  Methods for function \code{plot} in Package `solaR'
}

\section{Methods}{
  \describe{
    \item{\code{signature(x = "Shade")}}{display the results of
      the iteration with a level plot for the two-axis tracking, or with
      conventional plot for horizontal tracking and fixed systems.}
}}
\keyword{methods}
\author{Oscar Perpiñán Lamigueiro}
