\name{plot.sombra}
\Rdversion{1.1}
\encoding{UTF-8}
\alias{plot.sombra}

\title{
Plot the result of optimSombra
}
\description{
Plot the result of optimSombra
}
\usage{
\method{plot}{sombra}(x, titulo='', ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{object of class \code{sombra}, result of optimSombra}
  \item{titulo}{character, title of the plot}
  \item{\dots}{additional arguments to filled.contour or plot }
}
\references{
Perpiñán, O, Energía Solar Fotovoltaica, 2010.  (\url{http://procomun.wordpress.com/documentos/libroesf/})
}
\author{
Oscar Perpiñán Lamigueiro
}

\seealso{\code{\link{optimSombra}}}

\keyword{hplot}
