\name{E_prodEx}
\Rdversion{1.0}
\encoding{UTF-8}
\alias{prodEx}
\docType{data}
\title{Productivity of a set of PV systems of a PV plant.}

\description{A \code{zoo} object with the time evolution of the final productivity of a set
  of 22 systems of a large PV plant.}

\usage{data(prodEx)}

\references{O. Perpiñán, Statistical analysis of the performance and
  simulation of a two-axis tracking PV system, Solar Energy,
  83:11(2074–2085),
  2009.\url{http://oa.upm.es/1843/1/PERPINAN_ART2009_01.pdf}
}
\seealso{
  \code{\link{TargetDiagram}},
  \code{\link{analyzeData}}.
}

\keyword{datasets}
