\name{solaR.theme}
\alias{solaR.theme}
\docType{data}
\title{solaR theme}
\description{
A customized theme for lattice. It is based on the \code{custom.theme.2}
function of the \code{latticeExtra} package with the next values:
\itemize{

  \item{\code{pch=19}}
  \item{\code{cex=0.7}}
  \item{\code{region=rev(brewer.pal(9, 'YlOrRd'))}}
  \item{\code{strip.background$col='lightgray'}}
  \item{\code{strip.shingle$col='transparent'}}}
}

\details{

}

\keyword{datasets}
