\name{C_utils-angle}
\Rdversion{1.0}
\encoding{UTF-8}
\alias{d2r}
\alias{r2d}
\alias{h2r}
\alias{r2h}
\alias{r2sec}

\title{Conversion between angle units.}

\description{Several small functions to convert angle units. }

\usage{
d2r(x)
r2d(x)
h2r(x)
r2h(x)
r2sec(x)
}
\arguments{

  \item{x}{A numeric value.}
}

\value{A numeric value:
  \describe{
    \item{d2r: }{Degrees to radians.}
    \item{r2d: }{Radians to degrees.}
    \item{h2r: }{Hours to radians.}
    \item{r2h: }{Radians to hours.}
    \item{r2sec: }{Radians to seconds.}
  }
}

\author{Oscar Perpiñán Lamigueiro.}

\keyword{utilities}

