\name{solaR-defunct}
\encoding{UTF-8}
\alias{solaR-defunct}
\alias{readSIAR}
\alias{TargetDiagram}
\alias{analyzeData}

\title{Defunct functions in package \sQuote{solaR}}

\description{
  These functions are no longer available.
}

\details{
  \itemize{

    \item{\code{readSIAR}: The SIAR webpage cannot be accessed with a direct
      URL but using javascript code. Therefore, the function
      \code{readSIAR} no longer works. This help page is still here as a
      reference. The SIAR webpage is now \url{http://eportal.mapa.gob.es/websiar/Inicio.aspx}.}
    
    \item{\code{TargetDiagram}, \code{analyzeData}: Use the \code{tdr} package}

  }
  }
