\name{solaR-deprecated}
\encoding{UTF-8}
\alias{solaR-deprecated}
\alias{TargetDiagram}
\alias{analyzeData}

\title{Deprecated functions in package \sQuote{solaR}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{solaR} only, and will be defunct at the next release.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{TargetDiagram, analyzeData: Use the \code{tdr} package}

  }
  }
