\name{D_shadeplot-methods}
\docType{methods}
\alias{shadeplot}
\alias{shadeplot-methods}
\alias{shadeplot,Shade-method}
\title{Methods for Function shadeplot}
\description{
  Visualization of the content of a \code{\linkS4class{Shade}} object.
}

\section{Methods}{
  \describe{
    \item{\code{signature(x = "Shade")}}{display the results of
      the iteration with a level plot for the two-axis tracking, or with
      conventional plot for horizontal tracking and fixed systems.}
}}
\keyword{methods}
\author{Oscar Perpiñán Lamigueiro}
